@@ -4,16 +4,25 @@
 	"private": true,
 	"homepage": ".",
 	"dependencies": {
-		"axios": "^0.19.0",
+		"@date-io/moment": "^1.3.13",
+		"@material-ui/core": "^4.9.5",
+		"@material-ui/lab": "^4.0.0-alpha.45",
+		"@material-ui/pickers": "^3.2.10",
+		"axios": "^0.19.2",
 		"chart.js": "^2.9.3",
 		"moment": "^2.24.0",
 		"qs": "^6.9.1",
-		"react": "^16.12.0",
-		"react-chartjs-2": "^2.8.0",
-		"react-dom": "^16.12.0",
-		"react-icons": "^3.8.0",
+		"react": "^16.13.0",
+		"react-chartjs-2": "^2.9.0",
+		"react-dom": "^16.13.0",
+		"react-icons": "^3.9.0",
+		"react-promise-tracker": "^2.1.0",
+		"react-redux": "^7.2.0",
 		"react-router-dom": "^5.1.2",
-		"react-scripts": "^3.3.0"
+		"react-scripts": "^3.4.0",
+		"redux": "^4.0.5",
+		"redux-devtools-extension": "^2.13.8",
+		"redux-thunk": "^2.3.0"
 	},
 	"scripts": {
 		"start": "react-scripts start",
@@ -25,7 +34,15 @@
 		"extends": "react-app"
 	},
 	"browserslist": {
-		"production": [ ">0.2%", "not dead", "not op_mini all" ],
-		"development": [ "last 1 chrome version", "last 1 firefox version", "last 1 safari version" ]
+		"production": [
+			">0.2%",
+			"not dead",
+			"not op_mini all"
+		],
+		"development": [
+			"last 1 chrome version",
+			"last 1 firefox version",
+			"last 1 safari version"
+		]
 	}
 }
