@@ -1,13 +1,18 @@
 import React from 'react';
 import Box from './Box/Box';
 import classes from './Boxes.module.css';
+import { useSelector } from 'react-redux';
 
-const boxes = (props) => (
-    <div className={classes.Boxes}>
-        <Box type="temperature" value={props.temperature}/>
-        <Box type="humidity" value={props.humidity}/>
-        <Box type="brightness" value={props.brightness}/>
-    </div>
-)
+const Boxes = (props) => {
+	const values = useSelector((state) => state.data.values);
 
-export default boxes;
+	return (
+		<div className={classes.Boxes}>
+			<Box type="temperature" value={values.temperature[values.temperature.length - 1]} />
+			<Box type="humidity" value={values.humidity[values.humidity.length - 1]} />
+			<Box type="brightness" value={values.brightness[values.brightness.length - 1]} />
+		</div>
+	);
+};
+
+export default Boxes;
