@@ -8,16 +8,18 @@ const items = [
 	{ name: 'Knižnica', number: 70 },
 	{ name: 'Aj 1', number: 71 },
 	{ name: 'Nj 2', number: 77 },
-	{ name: 'Bio kabinet', number: 83 },
-	{ name: 'VI.OA', number: 84 }
+	{ name: 'Bio kabinet', number: 83 }
+	// { name: 'VI.OA', number: 84 }
 ];
 
 const sideDrawerItems = (props) => (
 	<React.Fragment>
-		<div>1. poschodie</div>
+		<div>Prízemie</div>
 		{items.map((item, index) => {
 			let header = null;
 			if (item.number === 29) {
+				header = '1. poschodie';
+			} else if (item.number === 53) {
 				header = '2. poschodie';
 			} else if (item.number === 77) {
 				header = '3. poschodie';
