@@ -1,8 +1,23 @@
 import React from 'react';
 import classes from './Backdrop.module.css';
+import { connect } from 'react-redux';
+import { toggleNav } from '../../../store/actions/index';
 
-const backdrop = (props) => (
-    props.show ? <div className={classes.Backdrop} onClick={props.clicked}></div> : null
-)
+const backdrop = (props) =>
+	props.isNavOpened ? (
+		<div className={classes.Backdrop} onClick={props.toggleNav.bind(this, props.isNavOpened)} />
+	) : null;
 
-export default backdrop;
+const mapStateToProps = (state) => {
+	return {
+		isNavOpened: state.ui.isNavOpened
+	};
+};
+
+const mapDispatchToProps = (dispatch) => {
+	return {
+		toggleNav: (isNavOpened) => dispatch(toggleNav(isNavOpened))
+	};
+};
+
+export default connect(mapStateToProps, mapDispatchToProps)(backdrop);
